+++ src/entities/Caixote/ModuloCozinha/Prateleiras.java
@@ -5,7 +5,7 @@ import entities.Caixote.Modulo;
 public class Prateleiras extends Modulo {
     // Implementar o que esta peça possuirá.
 
-    public Prateleiras(Double altura, Double largura, Double profundidade, Double espessura, Double espessuraFundo) {
+    public Prateleiras(Integer altura, Integer largura, Integer profundidade, Integer espessura, Integer espessuraFundo) {
         super(altura, largura, profundidade, espessura, espessuraFundo);
 
     }
