+++ src/entities/Caixote/ModuloCozinha/Rodape.java
@@ -5,7 +5,7 @@ import entities.Caixote.Modulo;
 public class Rodape extends Modulo {
     // Implementar o que esta peça possuirá.
 
-    public Rodape(Double altura, Double largura, Double profundidade, Double espessura, Double espessuraFundo) {
+    public Rodape(Integer altura, Integer largura, Integer profundidade, Integer espessura, Integer espessuraFundo) {
         super(altura, largura, profundidade, espessura, espessuraFundo);
 
     }
