+++ src/entities/Caixote/ModuloCozinha/Superior.java
@@ -5,7 +5,7 @@ import entities.Caixote.Modulo;
 public class Superior extends Modulo {
     // Implementar o que esta peça possuirá.
 
-    public Superior(Double altura, Double largura, Double profundidade, Double espessura, Double espessuraFundo) {
+    public Superior(Integer altura, Integer largura, Integer profundidade, Integer espessura, Integer espessuraFundo) {
         super(altura, largura, profundidade, espessura, espessuraFundo);
     }
